// src/app/api/model/contactDto.ts
/**
 * BandIT
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BandDto } from './bandDto';
import { ContactType } from './contactType';

export interface ContactDto { 
    id?: number;
    addedTimestamp?: Date;
    band?: BandDto;
    firstName?: string;
    contactType?: ContactType;
    lastName?: string;
    phoneNumber?: string;
    emailAddress?: string;
    description?: string;
}